refactor: split image preview into reusable editor
